+++ Acornote_Kotlin/app/src/main/java/tonnysunm/com/acornote/ui/colortag/ColorTagListAdapterHorizontal.kt
@@ -13,10 +13,10 @@ import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.activity_note.*
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import tonnysunm.com.acornote.ui.HomeActivity
 import tonnysunm.com.acornote.databinding.ListItemColortagHorizontalBinding
 import tonnysunm.com.acornote.databinding.ListItemEditBinding
 import tonnysunm.com.acornote.model.ColorTag
+import tonnysunm.com.acornote.ui.HomeActivity
 import tonnysunm.com.acornote.ui.note.NoteActivity
 import tonnysunm.com.acornote.ui.note.NoteFragment
 
@@ -25,7 +25,7 @@ private const val FooterType = 1
 
 
 class ColorTagListAdapterHorizontal(
-    var selectedColorTagId: Long?,
+    var selectedColorTagId: Int?,
     var array: List<ColorTag>
 ) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
